Fixing Issues
For real though, I got issues. Went through the issues on GitHub and
fixed all of them apart from clean up repeating styles and making sure
everything is in the correct place in my table of contents. I fixed some
of it but it is something that could be more successful towards the end
when I have everything styled. I will be doing my best to be organized
as I go though.