Assignment 7
Changes made from the origional design include:
Changing the hover state of the footer navigation because it was having
contrast issues. Made the about button on the navigation direct the user
to the about section rather than the process section because it makes
more sense. Added a hover state to the how to brew link so the user has
some sort of indication that it is clickable. Changed the position of
the logo on small display to give the buttons more room making it easy
for the user to press. I do believe that I am as far done as I can get.
And I am proud of myself for completing this website.